Update controller service

+++ eagle-api/src/main/java/com/alarm/eagle/api/domain/repository/LogRuleRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by luxiaoxun on 2020/01/29.
  */
-public interface LogRuleRepository extends JpaRepository<LogRuleDo, Integer> {
+public interface LogRuleRepository extends JpaRepository<LogRuleDo, Long> {
 }
